pyproject.toml
[tool.poetry]
name = "project-muserium"
version = "0.1.0"
description = "Django Backend dev"
authors = ["SigningOwl <whddms1208@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
boto3 = "^1.34.155"
django-storages = "^1.14.4"
django = "^5.0.8"
django-cors-headers = "^4.4.0"
uvicorn = "^0.30.5"
daphne = "^4.1.2"
channels = "^4.1.0"
django-environ = "^0.11.2"
pillow = "^10.4.0"
djangorestframework = "^3.15.2"
django-extensions = "^3.2.3"
werkzeug = "^3.0.3" # Not used in the production
djangorestframework-simplejwt = "^5.3.1"
dj-rest-auth = "^6.0.0"
django-allauth = "^64.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
